Mongolbank as an separate module [Refactor: no logic change]
